+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.3.0] - 2017-08-27
+### Added
+- Core: Core commands are not available in production by production. [#29](https://github.com/nunomaduro/laravel-zero/pull/29)
+- Core: Adds component:install command.
+- Core: Adds illuminate/database component.
+
 ## [3.2.0] - 2017-08-22
 ### Added
 - Core: Adds the executable bit on the compiled standalone phar on the builder command.